FastAPI/pyproject.toml
[tool.poetry]
name = "booking"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.31"
aiomysql = "^0.2.0"
pydantic = "^2.7.4"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"
aiomysql = "^0.2.0"
pytest-asyncio = "0.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
